fix: restaura Sushi com heranca apos merge

+++ src/main/java/org/example/Main.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class Main {
     public static void main(String[] args) {
-        Sushi s = new Sushi("Temaki Salmão", 5.50, 10);
+        Sushi s = new Sushi("Temaki Salmão", (int) 5.50, 10);
         s.exibirInfo();
         System.out.println("Total: R$ " + String.format("%.2f", s.calcularTotal()));
     }
